pyproject.toml
[tool.poetry]
name = "foodgpt"
version = "0.1.0"
description = "A food recommendator for HK and Singapore restaurants built on telegram"
authors = ["Reagan <rrr.chanhiulok@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.4"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
openai = "^1.37.1"
azure-ai-formrecognizer = "^3.3.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.5"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
